requirements.txt
aiohttp==0.21.5
chardet==2.3.0
python-decouple==3.0
